Simplify semantic color names
