+++ src/entities/language/plurals/PluralStrings.ts
@@ -20,15 +20,15 @@ export function getPluralRuleKeyLabel(key: PluralRuleKey): string {
 export function getPluralRuleColor(key: PluralRuleKey): string {
   switch (key) {
     case PluralRuleKey.Zero:
-      return 'var(--color-text-blue)';
+      return 'var(--color-blue)';
     case PluralRuleKey.One:
-      return 'var(--color-text-green)';
+      return 'var(--color-green)';
     case PluralRuleKey.Two:
-      return 'var(--color-text-yellow)';
+      return 'var(--color-yellow)';
     case PluralRuleKey.Few:
-      return 'var(--color-text-orange)';
+      return 'var(--color-orange)';
     case PluralRuleKey.Many:
-      return 'var(--color-text-red)';
+      return 'var(--color-red)';
     case PluralRuleKey.Other:
       return 'var(--color-button-secondary)';
   }
