+++ src/features/pagination/VisibleItemsMeter.tsx
@@ -107,7 +107,7 @@ const HighLimitWarning: React.FC<{ nShown: number }> = ({ nShown }) => {
 
   return (
     <div>
-      <TriangleAlertIcon size="1em" style={{ color: 'var(--color-text-yellow)' }} />
+      <TriangleAlertIcon size="1em" style={{ color: 'var(--color-yellow)' }} />
       There are <strong>{nShown.toLocaleString()}</strong> items visible, this may impact page
       performance. Consider reducing the limit to{' '}
       <HoverableButton
